+++ lib/plugins.dart
@@ -1,0 +1,3 @@
+library cruky.plugins_helper;
+
+export './src/scanner/scanner.dart' show PipelineMock;